+++ examples/resources/truenas_acme_dns_authenticator/resource.tf
@@ -2,7 +2,7 @@ resource "truenas_acme_dns_authenticator" "example" {
   name          = "cloudflare"
   authenticator = "cloudflare"
   attributes = {
-    cloudflare_email   = "dns@example.com"
-    cloudflare_api_key = "changeme"
+    cloudflare_email = "dns@example.com"
+    api_key          = "changeme"
   }
 }